services/digital-brain/core/tools/web_search.py
from .base import BaseTool
from typing import Dict, Any
import json
import requests
from loguru import logger

class WebSearchTool(BaseTool):
    name = "web_search"
    description = "Perform a real-time web search to get up-to-date information from the internet."
    
    def run(self, params: Dict[str, Any]) -> str:
        query = params.get("query")
        if not query:
            return "Error: Missing 'query' parameter."
            
        logger.info(f"[WebSearch] Searching for: {query}")
        
        # TODO: Replace with real SerpApi / Bing API Key
        # Currently using a mock response or a lightweight scraper if needed.
        # For production:
        # response = requests.get("https://serpapi.com/search", params={"q": query, "api_key": "YOUR_KEY"})
        
        # Mock Response for Demo
        mock_results = [
            {
                "title": f"Latest news about {query}",
                "snippet": f"Here is the most recent information regarding {query}. Market trends show positive growth...",
                "link": "https://example.com/news1"
            },
            {
                "title": f"{query} Analysis 2026",
                "snippet": f"A deep dive into {query} reveals significant changes in the landscape...",
                "link": "https://example.com/report"
            }
        ]
        
        return json.dumps({
            "query": query,
            "results": mock_results,
            "source": "MockSearchEngine (Set up API Key in Config)"
        })

    @property
    def schema(self):
        return {
            "name": self.name,
            "description": self.description,
            "parameters": {
                "type": "object",
                "properties": {
                    "query": {
                        "type": "string",
                        "description": "The search keywords or question."
                    }
                },
                "required": ["query"]
            }
        }
