services/digital-brain/requirements.txt
openai
pydantic
python-dotenv
requests
beautifulsoup4
zhipuai
dashscope
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
pypdf
tiktoken
pgvector
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psutil
python-pptx
